\name{lctools-package}
\alias{lctools-package}
\alias{lctools}
\docType{package}
\title{
Local multi-collinearity diagnostics and SIM tools}
\description{
The main purpose of lctools is to help testing the existence of local multi-collinearity among the explanatory variables of local regression models. The main function allows the computation of Local Pearson and Geographically Weighted Pearson Correlation Coefficients and tests for their significance. This package has also some other tools tha help computing variables for Spatial Interaction Models.  }

\details{
\tabular{ll}{
Package: \tab lctools\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2014-01-06\cr
License: \tab GPL (>= 2)
}
  
}
\author{
Stamatis Kalogirou

Maintainer: Stamatis Kalogirou <skalo@hua.gr}

\note{Acknowledgement: I am grateful to the University of Luxembourg and would like to personally thank Ass. Professor Geoffrey Caruso and Professor Christian Schulz for their support during my research visit at the Institute of Geography and Spatial Planning.  }

\references{
Kalogirou, S., 2011, Testing local versions of correlation coefficients, Review of Regional Research - Jahrbuch fur Regionalwissenschaft, 32, 1, pp. 45 - 61.

Kalogirou, S. (2013) Testing geographically weighted multicollinearity diagnostics, GISRUK 2013, Department of Geography and Planning, School of Environmental Sciences, University of Liverpool, Liverpool, UK, 3-5 April 2013.
\url{http://gisc.gr/docs/sk_papers/2_7_Kalogirou_2013.pdf}
}
\keyword{ package }